refactor: split repository interface
