+++ GyroData/Data/CoreMotionPlatform/Repository/CoreMotionRepository.swift
@@ -7,13 +7,6 @@
 
 import CoreMotion
 
-protocol CoreMotionRepositorable {
-    func startAccelerometer(_ completion: @escaping (MotionCoordinates) -> Void)
-    func startGyroscope(completion: @escaping (MotionCoordinates) -> Void)
-    func stopAccelerometer()
-    func stopGyroscope()
-}
-
 final class CoreMotionRepository: CoreMotionRepositorable {
     
     let dataSource: CoreMotionManager
